requirements.txt
altair==4.2.2
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
boltons==21.0.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
cheroot==9.0.0
click==8.1.7
click-spinner==0.1.10
colorama==0.4.6
colorlog==6.7.0
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dacite==1.8.1
datacommons==1.4.3
datacommons-pandas==0.0.3
datapane==0.16.7
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.2
django-tables2==2.6.0
djangorestframework==3.14.0
dominate==2.8.0
dulwich==0.21.5
entrypoints==0.4
executing==1.2.0
face==20.1.1
fastjsonschema==2.18.0
fonttools==4.41.1
furl==2.1.3
glom==23.3.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-resources==5.13.0
ipykernel==6.25.1
ipynbname==2023.2.0.0
ipython==8.14.0
jaraco.functools==3.8.1
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
micawber==0.5.5
mistune==3.0.1
monotonic==1.6
more-itertools==10.1.0
multimethod==1.9.1
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
numpy==1.25.1
orderedmultidict==1.0.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
posthog==2.5.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.7
pure-eval==0.2.2
pyarrow==10.0.1
pydantic==1.10.12
Pygments==2.16.1
pyngrok==5.2.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
stack-data==0.6.2
tabulate==0.9.0
tinycss2==1.2.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.4
vega-datasets==0.9.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.5.0
pandas==1.3.4

